move PagePlaceholder to components/
