+++ pages/_app.tsx
@@ -4,7 +4,7 @@ import '../src/components/Latex/LatexExample.css'
 import '../src/components/ScrollingText/ScrollingText.scss'
 import '../src/components/Marquee/Marquee.scss'
 import '../src/pages/Post/Post.css'
-import '../src/pages/PagePlaceholder/PagePlaceholder.css'
+import '../src/components/PagePlaceholder/PagePlaceholder.css'
 import '../src/components/Overlay/Overlay.scss'
 import '../src/components/PageLayout/components/MenuSeminars.css'
 import '../src/components/PageLayout/components/Footer.css'
