+++ pages/matik/tmm.tsx
@@ -1,7 +1,7 @@
 import {NextPage} from 'next'
 
 import {PageLayout} from '../../src/components/PageLayout/PageLayout'
-import {PagePlaceholder} from '../../src/pages/PagePlaceholder'
+import {PagePlaceholder} from '../../src/components/PagePlaceholder'
 
 const Tmm: NextPage = () => (
   <PageLayout seminarId={2}>
